% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_analysis.R
\name{copy_analysis}
\alias{copy_analysis}
\title{Copy Analysis Rmd and kihon.R to Working Directory}
\usage{
copy_analysis()
}
\description{
This function copies the analysis.Rmd and kihon.R files from the package to the current working directory.
}
